Use Structured Logging in a More Structured Way
In several places, empty curly braces (`{}`) were used in place of
structured logging properties. These were processed as literal
curly braces, and log context was lost. Where the value of the
property was an exception, an Exception-specialized overload was
chosen instead.

This did not alter any places where string interpolation was used
instead of structured logging properties. That could be work for
the future.